feat(illustrations): join admin-illustrations and admin-ui packages
