+++ packages/admin-ui/package.json
@@ -50,7 +50,6 @@
     "@emotion/babel-plugin": "^11.2.0",
     "@emotion/css": "^11.1.3",
     "@emotion/react": "^11.1.5",
-    "@vtex/admin-illustrations": "^0.1.3",
     "@vtex/admin-ui-core": "^0.4.0",
     "@vtex/admin-ui-hooks": "^0.1.3",
     "@vtex/admin-ui-react": "^0.4.1",
