+++ packages/admin-ui/src/components/Illustrations/components/Illustration.tsx
