+++ packages/admin-ui/src/components/Illustrations/index.ts
@@ -1,0 +1,4 @@
+export * from './components/Empty'
+export * from './components/Error'
+export * from './components/Illustration'
+export * from './components/NotFound'
