+++ packages/admin-ui/src/index.ts
@@ -1,6 +1,5 @@
 export * from '@vtex/admin-ui-core'
 export * from '@vtex/admin-ui-react'
-export * from '@vtex/admin-illustrations'
 export * from '@vtex/admin-ui-hooks'
 export * from '@vtex/admin-ui-util'
 export * from './components/Skeleton'
@@ -63,3 +62,4 @@ export * from './components/Grid'
 export * from './components/Field'
 export * from './components/Center'
 export * from './components/icons'
+export * from './components/Illustrations'
