--- packages/illustrations/CHANGELOG.md
@@ -1,25 +1,0 @@
-# Change Log
-
-All notable changes to this project will be documented in this file.
-See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
-
-## [0.1.3](https://github.com/vtex/onda/compare/@vtex/admin-illustrations@0.1.2...@vtex/admin-illustrations@0.1.3) (2021-12-13)
-
-**Note:** Version bump only for package @vtex/admin-illustrations
-
-## [0.1.2](https://github.com/vtex/onda/compare/@vtex/admin-illustrations@0.1.1...@vtex/admin-illustrations@0.1.2) (2021-07-28)
-
-**Note:** Version bump only for package @vtex/admin-illustrations
-
-## [0.1.1](https://github.com/vtex/onda/compare/@vtex/admin-illustrations@0.1.0...@vtex/admin-illustrations@0.1.1) (2021-04-27)
-
-### Bug Fixes
-
-- **table-empty-state:** collection empty illustration width and height ([bb1ba88](https://github.com/vtex/onda/commit/bb1ba88b96abeeb09ef792b1a298609a915972f1))
-- **table-empty-state:** up collection empty illustrtion ([18abb17](https://github.com/vtex/onda/commit/18abb17fcc4638b90b32c4e8fd62dca901f42e0d))
-
-# 0.1.0 (2021-04-20)
-
-### Features
-
-- **admin-illustrations:** new package to store illustrations ([27db803](https://github.com/vtex/onda/commit/27db803604190bf5aff5a999e7c26e4166abd397))
