--- packages/illustrations/README.md
@@ -1,3 +1,0 @@
-## admin illustrations
-
-[![NPM](https://img.shields.io/npm/v/@vtex/admin-illustrations.svg)](https://www.npmjs.com/package/@vtex/admin-illustrations)
