--- packages/illustrations/package.json
@@ -1,38 +1,0 @@
-{
-  "name": "@vtex/admin-illustrations",
-  "version": "0.1.3",
-  "main": "dist/index.js",
-  "module": "dist/admin-illustrations.esm.js",
-  "typings": "dist/index.d.ts",
-  "publishConfig": {
-    "access": "public",
-    "registry": "https://registry.npmjs.org"
-  },
-  "files": [
-    "dist",
-    "src"
-  ],
-  "engines": {
-    "node": ">=10"
-  },
-  "repository": {
-    "directory": "packages/illustrations",
-    "type": "git",
-    "url": "git+https://github.com/vtex/onda.git"
-  },
-  "bugs": {
-    "url": "https://github.com/vtex/onda/issues"
-  },
-  "scripts": {
-    "dev": "tsdx watch --verbose --noClean",
-    "build": "tsdx build --tsconfig tsconfig.json",
-    "build:fast": "tsdx build --tsconfig tsconfig.json --transpileOnly --format esm,cjs",
-    "test": "tsdx test --passWithNoTests",
-    "test:ci": "tsdx test --passWithNoTests --silent"
-  },
-  "peerDependencies": {
-    "react": ">=16",
-    "react-dom": ">=16"
-  },
-  "gitHead": "201e6a0149b4eb0b836a5f5f69b2497475877a71"
-}
