--- packages/illustrations/src/Collection/index.ts
@@ -1,3 +1,0 @@
-export * from './Empty'
-export * from './Error'
-export * from './NotFound'
