--- packages/illustrations/tsconfig.json
@@ -1,8 +1,0 @@
-{
-  "extends": "../../tsconfig.json",
-  "include": ["./src"],
-  "compilerOptions": {
-    "outDir": "./dist",
-    "resolveJsonModule": true
-  }
-}
